feat: extract receipt list screen as new file
